Divide models.py
