+++ opac/models/__init__.py
@@ -1,0 +1,1 @@
+from opac.models import masters, transactions
