+++ opac/models/abstracts.py
@@ -1,0 +1,15 @@
+from django.db import models
+
+
+class TimeStampedModel(models.Model):
+    class Meta:
+        abstract = True
+
+    created_at = models.DateTimeField(
+        '作成日時',
+        auto_now_add=True
+    )
+    updated_at = models.DateTimeField(
+        '更新日時',
+        auto_now=True
+    )
