+++ opac/models/masters/__init__.py
@@ -1,0 +1,6 @@
+from opac.models.masters.author import Author
+from opac.models.masters.book import Book
+from opac.models.masters.library import Library
+from opac.models.masters.publisher import Publisher
+from opac.models.masters.stock import Stock
+from opac.models.masters.translator import Translator
