+++ opac/models/transactions/__init__.py
@@ -1,0 +1,4 @@
+from opac.models.transactions.holding import Holding
+from opac.models.transactions.lending import Lending
+from opac.models.transactions.renewing import Renewing
+from opac.models.transactions.reservation import Reservation
